chore: consolidate 31 duplicated formatting helpers into shared term-format module
Extract padRight (14 copies), truncate (4), formatDate (3), formatTimestamp (6),
and formatTime (3) into src/lib/term-format.ts. Replaces 213 lines of duplicated
code across 16 CLI command files with single-line imports.
